feat: move notify-client2 library to fido2 project #5030
