+++ jans-fido2/notify-client2/src/main/java/io/jans/notify/model/RegisterDeviceResponse.java
