--- jans-notify/README.md
@@ -1,2 +1,0 @@
-# Janssen Notify
-Janssen Notify is a simple push notification server.
