--- jans-notify/client/pom.xml
@@ -1,32 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-
-	<modelVersion>4.0.0</modelVersion>
-	<artifactId>jans-notify-client</artifactId>
-	<name>notify Client</name>
-	<packaging>jar</packaging>
-
-	<parent>
-		<artifactId>jans-notify-parent</artifactId>
-		<groupId>io.jans</groupId>
-		<version>1.0.14-SNAPSHOT</version>
-	</parent>
-
-	<dependencies>
-		<!-- jans -->
-		<dependency>
-			<groupId>io.jans</groupId>
-			<artifactId>jans-notify-model</artifactId>
-		</dependency>
-
-		<!-- Resteasy -->
-		<dependency>
-			<groupId>org.jboss.resteasy</groupId>
-			<artifactId>resteasy-client</artifactId>
-		</dependency>
-	   <dependency>
-	        <groupId>org.jboss.resteasy</groupId>
-	        <artifactId>resteasy-jackson2-provider</artifactId>
-	    </dependency>
-	</dependencies>
-</project>
