--- jans-notify/client/src/test/java/org/gluu/oxbroker/client/.gitignore
@@ -1,1 +1,0 @@
-/dev/
