--- jans-notify/model/pom.xml
@@ -1,21 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-
-	<modelVersion>4.0.0</modelVersion>
-	<artifactId>jans-notify-model</artifactId>
-	<name>notify Model</name>
-	<packaging>jar</packaging>
-
-	<parent>
-		<artifactId>jans-notify-parent</artifactId>
-		<groupId>io.jans</groupId>
-		<version>1.0.14-SNAPSHOT</version>
-	</parent>
-
-	<dependencies>
-		<dependency>
-			<groupId>com.fasterxml.jackson.core</groupId>
-			<artifactId>jackson-annotations</artifactId>
-		</dependency>
-	</dependencies>
-</project>
