--- jans-notify/server/conf/oxnotify-access.json
@@ -1,18 +1,0 @@
-{
-   "clientConfigurations":[
-      {
-         "clientId":"ce_gcm_client",
-         "platformId":"ce_gcm",
-         "accessKeyId":"client_access_key_id",
-         "secretAccessKey":"encoded_client_secret_access_key",
-         "enabled":true
-      },
-      {
-         "clientId":"ce_apns_client",
-         "platformId":"ce_apns",
-         "accessKeyId":"client_access_key_id",
-         "secretAccessKey":"encoded_client_secret_access_key",
-         "enabled":true
-      }
-   ]
-}
