--- jans-notify/server/conf/oxnotify-config.json
@@ -1,24 +1,0 @@
-{
-   "issuer":"https://demo.gluu.org",
-   "baseEndpoint":"https://demo.gluu.org/oxnotify",
-   "platformConfigurations":[
-      {
-         "platformId":"ce_gcm",
-         "platform":"GCM",
-         "platformArn":"arn_to_gcm_enpoint",
-         "accessKeyId":"access_key_id",
-         "secretAccessKey":"encoded_secret_access_key",
-         "region":"region",
-         "enabled":true
-      },
-      {
-         "platformId":"ce_apns",
-         "platform":"APNS",
-         "platformArn":"arn_to_apns_enpoint",
-         "accessKeyId":"access_key_id",
-         "secretAccessKey":"encoded_secret_access_key",
-         "region":"region",
-         "enabled":true
-      }
-   ]
-}
