--- jans-notify/server/src/main/java/io/jans/notify/.gitignore
@@ -1,1 +1,0 @@
-/dev/
