--- jans-notify/server/src/main/java/io/jans/notify/model/conf/ClientConfiguration.java
@@ -1,66 +1,0 @@
-/*
- * Janssen Project software is available under the MIT License (2008). See http://opensource.org/licenses/MIT for full text.
- *
- * Copyright (c) 2020, Janssen Project
- */
-
-package io.jans.notify.model.conf;
-
-import jakarta.enterprise.inject.Vetoed;
-
-/**
- * @author Yuriy Movchan
- * @version September 15, 2017
- */
-@Vetoed
-public class ClientConfiguration {
-
-	private String clientId;
-	private String platformId;
-	
-	private String accessKeyId;
-	private String secretAccessKey;
-
-	private boolean enabled;
-
-	public String getClientId() {
-		return clientId;
-	}
-
-	public void setClientId(String clientId) {
-		this.clientId = clientId;
-	}
-
-	public String getPlatformId() {
-		return platformId;
-	}
-
-	public void setPlatformId(String platformId) {
-		this.platformId = platformId;
-	}
-
-	public String getAccessKeyId() {
-		return accessKeyId;
-	}
-
-	public void setAccessKeyId(String accessKeyId) {
-		this.accessKeyId = accessKeyId;
-	}
-
-	public String getSecretAccessKey() {
-		return secretAccessKey;
-	}
-
-	public void setSecretAccessKey(String secretAccessKey) {
-		this.secretAccessKey = secretAccessKey;
-	}
-
-	public boolean isEnabled() {
-		return enabled;
-	}
-
-	public void setEnabled(boolean enabled) {
-		this.enabled = enabled;
-	}
-
-}
