--- jans-notify/server/src/main/java/io/jans/notify/service/HealthCheckController.java
@@ -1,34 +1,0 @@
-/*
- * Janssen Project software is available under the MIT License (2008). See http://opensource.org/licenses/MIT for full text.
- *
- * Copyright (c) 2020, Janssen Project
- */
-
-package io.jans.notify.service;
-
-import jakarta.enterprise.context.ApplicationScoped;
-import jakarta.ws.rs.GET;
-import jakarta.ws.rs.POST;
-import jakarta.ws.rs.Path;
-import jakarta.ws.rs.Produces;
-import jakarta.ws.rs.core.MediaType;
-
-/**
- * Health check controller
- * 
- * @author Yuriy Movchan
- * @version Jul 24, 2020
- */
-@ApplicationScoped
-@Path("/")
-public class HealthCheckController {
-
-    @GET
-    @POST
-    @Path("/health-check")
-    @Produces(MediaType.APPLICATION_JSON)
-	public String healthCheckController() {
-        return "{\"status\":\"running\"}";
-	}
-
-}
