--- jans-notify/server/src/main/java/io/jans/notify/service/NetworkService.java
@@ -1,52 +1,0 @@
-/*
- * Janssen Project software is available under the MIT License (2008). See http://opensource.org/licenses/MIT for full text.
- *
- * Copyright (c) 2020, Janssen Project
- */
-
-package io.jans.notify.service;
-
-import jakarta.enterprise.context.ApplicationScoped;
-import jakarta.inject.Inject;
-import jakarta.inject.Named;
-import jakarta.servlet.http.HttpServletRequest;
-
-import org.slf4j.Logger;
-
-/**
- * @author Yuriy Movchan
- * @version October 10, 2017
- */
-@ApplicationScoped
-@Named
-public class NetworkService {
-
-	@Inject
-	private Logger log;
-
-
-    public String getIpAddress(HttpServletRequest httpRequest) {
-        final String[] HEADERS_TO_TRY = {
-                "X-Forwarded-For",
-                "Proxy-Client-IP",
-                "WL-Proxy-Client-IP",
-                "HTTP_X_FORWARDED_FOR",
-                "HTTP_X_FORWARDED",
-                "HTTP_X_CLUSTER_CLIENT_IP",
-                "HTTP_CLIENT_IP",
-                "HTTP_FORWARDED_FOR",
-                "HTTP_FORWARDED",
-                "HTTP_VIA",
-                "REMOTE_ADDR"
-        };
-        for (String header : HEADERS_TO_TRY) {
-            String ip = httpRequest.getHeader(header);
-            if (ip != null && ip.length() != 0 && !"unknown".equalsIgnoreCase(ip)) {
-                return ip;
-            }
-        }
-
-        return httpRequest.getRemoteAddr();
-    }
-
-}
