--- jans-notify/server/src/main/resources/META-INF/beans.xml
@@ -1,6 +1,0 @@
-<!-- Marker file indicating CDI should be enabled -->
-<beans xmlns="https://jakarta.ee/xml/ns/jakartaee"
-        xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-        xsi:schemaLocation="https://jakarta.ee/xml/ns/jakartaee https://jakarta.ee/xml/ns/jakartaee/beans_3_0.xsd"
-        bean-discovery-mode="all" version="3.0">
-</beans>
