--- jans-notify/server/src/main/resources/log4j2.xml
@@ -1,34 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<Configuration packages="org.gluu.log">
-	<Appenders>
-		<Console name="STDOUT" target="SYSTEM_OUT">
-			<PatternLayout pattern="%d %-5p [%t] [%C{6}] (%F:%L) - %m%n" />
-		</Console>
-
-		<RollingFile name="FILE" fileName="${sys:log.base}/logs/oxnotify.log" filePattern="${sys:log.base}/logs/oxnotify-%d{yyyy-MM-dd}-%i.log">
-
-			<PatternLayout pattern="%d %-5p [%t] [%C{6}] (%F:%L) - %m%n" />
-
-			<Policies>
-				<TimeBasedTriggeringPolicy interval="1" modulate="true" />
-				<SizeBasedTriggeringPolicy size="250 MB" />
-			</Policies>
-			<DefaultRolloverStrategy max="30" />
-		</RollingFile>
-	</Appenders>
-
-	<Loggers>
-		<!-- ############### oxnotify ################# -->
-		<Logger name="org.gluu.oxnotify" level="${log4j.default.log.level}" />
-		<Logger name="org.gluu.service" level="${log4j.default.log.level}" />
-
-		<!-- ############### Gluu ################# -->
-		<Logger name="org.gluu" level="${log4j.default.log.level}" />
-
-		<Root level="info">
-			<AppenderRef ref="FILE" />
-			<AppenderRef ref="STDOUT" />
-		</Root>
-	</Loggers>
-</Configuration>
