--- jans-notify/server/src/main/webapp-jetty/WEB-INF/jetty-env.xml
@@ -1,22 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE Configure PUBLIC "-//Jetty//Configure//EN" "https://www.eclipse.org/jetty/configure_10_0.dtd">
-
-<Configure id="webAppCtx"
-	class="org.eclipse.jetty.webapp.WebAppContext">
-	<New id="BeanManager" class="org.eclipse.jetty.plus.jndi.Resource">
-		<Arg>
-			<Ref refid="webAppCtx" />
-		</Arg>
-
-		<Arg>BeanManager</Arg>
-
-		<Arg>
-			<New class="javax.naming.Reference">
-				<Arg>jakarta.enterprise.inject.spi.BeanManager</Arg>
-				<Arg>org.jboss.weld.resources.ManagerObjectFactory</Arg>
-				<Arg />
-			</New>
-		</Arg>
-	</New>
-
-</Configure>
