--- jans-notify/server/src/main/webapp-jetty/WEB-INF/jetty-web.xml
@@ -1,30 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE Configure PUBLIC "-//Jetty//Configure//EN" "https://www.eclipse.org/jetty/configure_10_0.dtd">
-
-<Configure class="org.eclipse.jetty.webapp.WebAppContext">
-
-	<Get name="serverClassMatcher">
-		<Call name="add">
-			<Arg>-org.eclipse.jetty.util.Decorator</Arg>
-		</Call>
-	
-		<Call name="add">
-			<Arg>-org.eclipse.jetty.util.DecoratedObjectFactory</Arg>
-		</Call>
-	
-		<Call name="add">
-			<Arg>-org.eclipse.jetty.server.handler.ContextHandler.</Arg>
-		</Call>
-	
-		<Call name="add">
-			<Arg>-org.eclipse.jetty.server.handler.ContextHandler</Arg>
-		</Call>
-	
-		<Call name="add">
-			<Arg>-org.eclipse.jetty.servlet.ServletContextHandler</Arg>
-		</Call>
-	</Get>
-
-	<Set name="maxFormContentSize" type="int">50000000</Set>
-
-</Configure>
