--- jans-notify/server/src/main/webapp-jetty/WEB-INF/web.xml
@@ -1,33 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<web-app version="3.1"
-	xmlns="http://xmlns.jcp.org/xml/ns/javaee/"
-	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-	xsi:schemaLocation="http://xmlns.jcp.org/xml/ns/javaee http://xmlns.jcp.org/xml/ns/javaee/web-app_3_1.xsd">
-
-	<display-name>oxNotify Server</display-name>
-
-	<!-- Jetty -->
-	<context-param>
-		<param-name>org.eclipse.jetty.servlet.Default.dirAllowed</param-name>
-		<param-value>false</param-value>
-	</context-param>
-    
-	<!-- Weld -->
-	<context-param>
-		<param-name>org.jboss.weld.development</param-name>
-		<param-value>false</param-value>
-	</context-param>
-
-	<!-- Weld -->
-	<listener>
-		<listener-class>org.jboss.weld.environment.servlet.Listener</listener-class>
-	</listener>
-
-	<resource-env-ref>
-		<description>Object factory for the CDI Bean Manager</description>
-		<resource-env-ref-name>BeanManager</resource-env-ref-name>
-		<resource-env-ref-type>jakarta.enterprise.inject.spi.BeanManager</resource-env-ref-type>
-	</resource-env-ref>
-
-</web-app>
