--- jans-notify/server/src/main/webapp/META-INF/context.xml
@@ -1,2 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<Context antiJARLocking="true" path="/oxnotify"/>
