--- jans-notify/server/src/main/webapp/index.jsp
@@ -1,15 +1,0 @@
-<%--
-  Author: Yuriy Movchan
-  Version: September 15, 2017
---%>
-<%@ page contentType="text/html;charset=UTF-8" language="java" %>
-<html>
-<head>
-    <title>oxNotify</title>
-</head>
-<body>
-<h1>
-    oxNotify
-</h1>
-</body>
-</html>
